Supervised Method
Supervised Mining:

Each team is asked, after the exploratory analysis, to perform a supervised classification to classify the handwritten digits. The aim is to predict the CLASS variable (you should have already created) which values are 0, 1, ... or 9 corresponding to the handwritten digits shown in the image. The complete data set contains a balanced number of handwritten digits. These proportions must be observed in both the training set and the test set.